replace duplicate typing functions in operation.ts and replace duplicated naming function in ProtocolGraphQLConfiguration
